requirements.txt
asgiref==3.4.1
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.6
cryptography==3.4.8
defusedxml==0.7.1
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.2.7
django-allauth==0.45.0
django-toolbelt==0.0.1
djangorestframework==3.12.4
fontawesome-free==5.15.4
gunicorn==20.1.0
idna==3.2
oauthlib==3.1.1
psycopg2-binary==2.9.1
pycparser==2.20
PyJWT==2.1.0
python-dotenv==0.19.1
python3-openid==3.2.0
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
sqlparse==0.4.1
static3==0.7.0
urllib3==1.26.7
